fix: isolate mock.module test to prevent cross-file pollution
- Extract getTemplateVersion fallback test to separate file
  (Bun mock.module leaks globally; file-level isolation is the only fix)
- Remove mock import from main projects.test.ts
- Fix Biome line length violations in test file
